+++ internal/command/postgres/db.go
@@ -87,7 +87,7 @@ func runListDbs(ctx context.Context) error {
 	var leaderIp string
 	switch app.PlatformVersion {
 	case "nomad":
-		if err := NomadPGVersionCompatible(app, MinPostgresHaVersion, MinPostgresHaVersion); err != nil {
+		if err := nomadVersionCompatible(app, MinPostgresHaVersion, MinPostgresHaVersion); err != nil {
 			return err
 		}
 		pgInstances, err := agentclient.Instances(ctx, app.Organization.Slug, app.Name)
@@ -111,13 +111,13 @@ func runListDbs(ctx context.Context) error {
 		if err != nil {
 			return fmt.Errorf("machines could not be retrieved %w", err)
 		}
-		if err := MachinePGVersionCompatible(members, MinPostgresHaVersion, MinPostgresHaVersion); err != nil {
+		if err := machineVersionCompatible(members, MinPostgresHaVersion, MinPostgresHaVersion); err != nil {
 			return err
 		}
 		if len(members) == 0 {
 			return fmt.Errorf("no 6pn ips founds for %s app", app.Name)
 		}
-		leader, _ := MachinesNodeRoles(ctx, members)
+		leader, _ := machineNodeRoles(ctx, members)
 		leaderIp = leader.PrivateIP
 	default:
 		return fmt.Errorf("unsupported platform %s", app.PlatformVersion)
