+++ internal/command/restart/restart_machines.go
@@ -4,18 +4,14 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/superfly/flyctl/agent"
 	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/flaps"
-	"github.com/superfly/flyctl/flypg"
 	"github.com/superfly/flyctl/internal/command/machine"
 	"github.com/superfly/flyctl/internal/command/postgres"
-	"github.com/superfly/flyctl/internal/flag"
 	"github.com/superfly/flyctl/internal/watch"
-	"github.com/superfly/flyctl/iostreams"
 )
 
-func runMachineRestart(ctx context.Context, app *api.AppCompact) error {
+func RunMachineRestart(ctx context.Context, app *api.AppCompact) error {
 	flapsClient := flaps.FromContext(ctx)
 
 	machines, err := flapsClient.ListActive(ctx)
@@ -41,7 +37,7 @@ func runMachineRestart(ctx context.Context, app *api.AppCompact) error {
 	}
 
 	if app.PostgresAppRole.Name == "postgres_cluster" {
-		return runMachinePGRestart(ctx, machines)
+		return postgres.RestartMachines(ctx, machines)
 	}
 
 	return runMachineBaseRestart(ctx, machines)
@@ -59,77 +55,4 @@ func runMachineBaseRestart(ctx context.Context, machines []*api.Machine) error {
 	}
 
 	return nil
-}
-
-func runMachinePGRestart(ctx context.Context, machines []*api.Machine) (err error) {
-	var (
-		io                   = iostreams.FromContext(ctx)
-		colorize             = io.ColorScheme()
-		dialer               = agent.DialerFromContext(ctx)
-		MinPostgresHaVersion = "0.0.20"
-	)
-
-	if err := postgres.MachinePGVersionCompatible(machines, MinPostgresHaVersion, MinPostgresHaVersion); err != nil {
-		return err
-	}
-
-	leader, replicas := postgres.MachinesNodeRoles(ctx, machines)
-
-	// unless flag.force is set, we should error if leader==nil
-	if flag.GetBool(ctx, "force") && leader == nil {
-		fmt.Fprintln(io.Out, colorize.Yellow("No leader found, but continuing with restart"))
-	} else if leader == nil {
-		return fmt.Errorf("no active leader found")
-	}
-
-	fmt.Fprintln(io.Out, "Identifying cluster role(s)")
-
-	for _, machine := range machines {
-		fmt.Fprintf(io.Out, "  Machine %s: %s\n", colorize.Bold(machine.ID), postgres.MachineRole(machine))
-	}
-
-	if len(replicas) > 0 {
-		for _, replica := range replicas {
-			fmt.Fprintf(io.Out, "Restarting machine %s\n", colorize.Bold(replica.ID))
-
-			if err = machine.Restart(ctx, replica.ID, "", 120, false); err != nil {
-				return err
-			}
-			// wait for health checks to pass
-			if err := watch.MachinesChecks(ctx, []*api.Machine{replica}); err != nil {
-				return fmt.Errorf("failed to wait for health checks to pass: %w", err)
-			}
-		}
-	}
-
-	// Don't attempt to failover unless we have in-region replicas
-	inRegionReplicas := 0
-	for _, replica := range replicas {
-		if replica.Region == leader.Region {
-			inRegionReplicas++
-		}
-	}
-
-	if inRegionReplicas > 0 {
-		pgclient := flypg.NewFromInstance(leader.PrivateIP, dialer)
-
-		fmt.Fprintf(io.Out, "Attempting to failover %s\n", colorize.Bold(leader.ID))
-		if err := pgclient.Failover(ctx); err != nil {
-			fmt.Fprintln(io.Out, colorize.Red(fmt.Sprintf("failed to perform failover: %s", err.Error())))
-		}
-	}
-
-	fmt.Fprintf(io.Out, "Restarting machine %s\n", colorize.Bold(leader.ID))
-
-	if err = machine.Restart(ctx, leader.ID, "", 120, false); err != nil {
-		return err
-	}
-	// wait for health checks to pass
-	if err := watch.MachinesChecks(ctx, []*api.Machine{leader}); err != nil {
-		return fmt.Errorf("failed to wait for health checks to pass: %w", err)
-	}
-
-	fmt.Fprintf(io.Out, "Postgres cluster has been successfully restarted!\n")
-
-	return
 }
